requirements.txt
bcrypt>=4.0.0
cryptography>=42.0.0
google-genai>=0.3.0
numpy
openai>=1.50.0
pandas>=2.0.0
pymupdf>=1.25.0
python-dotenv
pytest
requests
streamlit>=1.41.0
tenacity>=8.2.0
extra-streamlit-components>=0.1.71
